requirements.txt
beautifulsoup4==4.8.2
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.11
django-bootstrap4==1.1.1
django-datatable-view==0.9.0
django-heroku==0.3.1
django-tables2==2.2.1
django-tempus-dominus==5.1.2.13
django-xhtml2pdf==0.0.4
django.js==0.8.1
gunicorn==20.0.4
html5lib==1.0.1
Pillow==7.0.0
psycopg2==2.8.5
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.42
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
